Major refactoring
- Re-structure to follow a standar python package structure
- Add __init__.py
- Add __main__.py as a CLI entry point
- Rename main.py to prolink.py
- Remove unused import and sort the rest
- Read the 'parameters.cfg' natively and use them automatically
- Change the parameters to be a plain dict instead of a dict of dicts
