+++ ProLink/modules/blast.py
@@ -1,24 +1,24 @@
+
+import os
+from io import StringIO
+
 from Bio import SeqIO
+from Bio.Blast import NCBIWWW, NCBIXML
 from Bio.Seq import Seq
 from Bio.SeqRecord import SeqRecord
-from Bio.Blast import NCBIWWW
-from Bio.Blast import NCBIXML
-from io import StringIO
-import subprocess
-import os
+
 
 def blast(hitlist_range, blast_database, blast_filename,  my_seq_record):
     print("----Searching in BLAST----")
     print("Hitlist size:    " + str(hitlist_range))
     print("Database:        " + blast_database)
     print("Output filename: " + blast_filename)
-    
     result_handle = NCBIWWW.qblast("blastp", str(blast_database), my_seq_record.seq, hitlist_size=hitlist_range)
     print(result_handle)
-
-    with open(blast_filename, 'w') as save_file: 
+    with open(blast_filename, 'w') as save_file:
         blast_results = result_handle.read()
         save_file.write(blast_results)
+
 
 def parse(my_sequence_index, hitlist_range, blast_filename, remove_gaps, expected_min_identity, outputs_dir):
     with open(blast_filename,"r") as f:
@@ -34,7 +34,7 @@ def parse(my_sequence_index, hitlist_range, blast_filename, remove_gaps, expecte
                         ),
                         id=alignment.title,
                     )
-            print('>', alignment.title) 
+            print('>', alignment.title)
             print(hsp.sbjct)
             if (hsp.identities / alignment.length) < expected_min_identity:
                     low_identity_seqs += 1
@@ -48,6 +48,7 @@ def parse(my_sequence_index, hitlist_range, blast_filename, remove_gaps, expecte
     found_sequences_fastafile= "./" + outputs_dir + "/protein_" + str(my_sequence_index) + "/found_sequences.fasta"
     SeqIO.write(found_sequences, found_sequences_fastafile, "fasta")
     return low_identity_seqs
+
 
 def p_parse(my_sequence_index, hitlist_range, blast_filename, remove_gaps, expected_min_identity, low_identity_seqs, max_low_identity_seqs, outputs_dir):
     with open(blast_filename,"r") as f:
@@ -65,11 +66,11 @@ def p_parse(my_sequence_index, hitlist_range, blast_filename, remove_gaps, expec
                           ),
                           id=alignment.title,
                       )
-              sequence_index += 1 
+              sequence_index += 1
               print("Sequence num " + str(sequence_index))
-              print('>', alignment.title) 
+              print('>', alignment.title)
               print(hsp.sbjct)
-              
+
               if (hsp.identities / alignment.length) < expected_min_identity:
                   low_identity_seqs += 1
                   print("low identity seq!")
@@ -86,10 +87,11 @@ def p_parse(my_sequence_index, hitlist_range, blast_filename, remove_gaps, expec
     found_sequences_fastafile= "./" + outputs_dir + "/protein_" + str(my_sequence_index) + "/found_sequences.fasta"
     SeqIO.write(found_sequences, found_sequences_fastafile, "fasta")
     return low_identity_seqs
-  
-def p_blast(my_sequence_index, blast_database, hitlist_range, my_seq_record, blast_filename, found_sequences_fastafile, remove_gaps, expected_min_identity, min_low_identity_seqs, max_low_identity_seqs, additional_hits, outputs_dir):  
+
+
+def p_blast(my_sequence_index, blast_database, hitlist_range, my_seq_record, blast_filename, found_sequences_fastafile, remove_gaps, expected_min_identity, min_low_identity_seqs, max_low_identity_seqs, additional_hits, outputs_dir):
     blast(hitlist_range, blast_database, blast_filename,  my_seq_record)
-    
+
     if parse(my_sequence_index, hitlist_range, blast_filename, remove_gaps, expected_min_identity, outputs_dir) < min_low_identity_seqs:
         print()
         print("The number of low identity sequences is below the desired value")
@@ -109,6 +111,4 @@ def p_blast(my_sequence_index, blast_database, hitlist_range, my_seq_record, bla
             blast(hitlist_range, blast_database, blast_filename,  my_seq_record)
             low_identity_seqs = 0
             p_parse(my_sequence_index, hitlist_range, blast_filename, remove_gaps, expected_min_identity, low_identity_seqs, max_low_identity_seqs, outputs_dir)
-    print()
-    print("Pro blast done succesfully")
-    print()
+    print("\nPro blast done succesfully\n")
